+++ v2/backend/application/credentials/ports.py
@@ -2,5 +2,7 @@ from typing import Protocol
 
 
 class CredentialsPort(Protocol):
-    def save(self, provider: str, api_key: str, secret_token: str, api_url: str) -> None: ...
+    def save(
+        self, provider: str, api_key: str, secret_token: str, api_url: str
+    ) -> None: ...
     def load(self, provider: str) -> dict | None: ...
