+++ v2/backend/application/portfolio/factory.py
@@ -3,8 +3,12 @@ from dotenv import load_dotenv
 from shared.utils.custom_logger import customer_logger
 from backend.application.portfolio.service import PortfolioService
 from backend.infrastructure.portfolio.repository_factory import RepositoryFactory
-from backend.infrastructure.portfolio.asset_analytics_repository import AssetAnalyticsRepository
-from backend.infrastructure.portfolio.portfolio_query_repository import PortfolioQueryRepository
+from backend.infrastructure.portfolio.asset_analytics_repository import (
+    AssetAnalyticsRepository,
+)
+from backend.infrastructure.portfolio.portfolio_query_repository import (
+    PortfolioQueryRepository,
+)
 
 logging = customer_logger("Portfolio Service")
 load_dotenv()
