+++ v2/backend/application/rebalancing/factory.py
@@ -1,6 +1,10 @@
 from backend.application.rebalancing.service import RebalancingService
-from backend.infrastructure.rebalancing.rebalance_config_repository import PostgresRebalanceConfigRepository
-from backend.infrastructure.rebalancing.rebalance_plan_repository import PostgresRebalancePlanRepository
+from backend.infrastructure.rebalancing.rebalance_config_repository import (
+    PostgresRebalanceConfigRepository,
+)
+from backend.infrastructure.rebalancing.rebalance_plan_repository import (
+    PostgresRebalancePlanRepository,
+)
 
 
 def build_rebalancing_service() -> RebalancingService:
