+++ v2/frontend/src/App.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom'
 import { ThemeProvider, CssBaseline } from '@mui/material'
 import { buildTheme } from './theme/theme'
 import { useAppStore } from './store/useAppStore'
-import AppShell from './pages/portfolio/AppShell'
+import AppShell from './layouts/AppShell'
 import PortfolioTab from './pages/portfolio/tabs/PortfolioTab'
 import RiskTab from './pages/portfolio/tabs/RiskTab'
 import OpportunitiesTab from './pages/portfolio/tabs/OpportunitiesTab'
