+++ v2/frontend/src/components/charts/asset/AssetPnlChart.tsx
