+++ v2/frontend/src/components/charts/asset/AssetPriceChart.tsx
