+++ v2/frontend/src/components/charts/asset/AssetReturnChart.tsx
