+++ v2/frontend/src/components/charts/asset/AssetValueChart.tsx
