+++ v2/frontend/src/components/charts/asset/index.ts
@@ -1,0 +1,4 @@
+export { default as AssetPnlChart } from './AssetPnlChart'
+export { default as AssetPriceChart } from './AssetPriceChart'
+export { default as AssetReturnChart } from './AssetReturnChart'
+export { default as AssetValueChart } from './AssetValueChart'
