+++ v2/frontend/src/components/charts/portfolio/DrawdownChart.tsx
