+++ v2/frontend/src/components/charts/portfolio/PortfolioPnlChart.tsx
