+++ v2/frontend/src/components/charts/portfolio/PortfolioValueChart.tsx
