+++ v2/frontend/src/components/charts/portfolio/index.ts
@@ -1,0 +1,3 @@
+export { default as PortfolioValueChart } from './PortfolioValueChart'
+export { default as PortfolioPnlChart } from './PortfolioPnlChart'
+export { default as DrawdownChart } from './DrawdownChart'
