+++ v2/frontend/src/components/charts/position/PositionPerformanceMap.tsx
