+++ v2/frontend/src/components/charts/position/PositionWeightChart.tsx
