+++ v2/frontend/src/components/charts/position/VarByPositionChart.tsx
