+++ v2/frontend/src/components/charts/position/index.ts
@@ -1,0 +1,3 @@
+export { default as PositionWeightChart } from './PositionWeightChart'
+export { default as PositionPerformanceMap } from './PositionPerformanceMap'
+export { default as VarByPositionChart } from './VarByPositionChart'
