+++ v2/frontend/src/components/charts/screener/DailyMoversTable.tsx
