+++ v2/frontend/src/components/charts/screener/LosersChart.tsx
