+++ v2/frontend/src/components/charts/screener/OpportunitiesChart.tsx
