+++ v2/frontend/src/components/charts/screener/UnprofitablePnlChart.tsx
