+++ v2/frontend/src/components/charts/screener/WinnersChart.tsx
