+++ v2/frontend/src/components/charts/screener/index.ts
@@ -1,0 +1,5 @@
+export { default as WinnersChart } from './WinnersChart'
+export { default as LosersChart } from './LosersChart'
+export { default as OpportunitiesChart } from './OpportunitiesChart'
+export { default as UnprofitablePnlChart } from './UnprofitablePnlChart'
+export { default as DailyMoversTable } from './DailyMoversTable'
