+++ v2/frontend/src/components/charts/shared/ProfitabilityChart.tsx
