+++ v2/frontend/src/components/charts/shared/SparklineChart.tsx
