+++ v2/frontend/src/components/charts/shared/index.ts
@@ -1,0 +1,2 @@
+export { default as SparklineChart } from './SparklineChart'
+export { default as ProfitabilityChart } from './ProfitabilityChart'
