+++ v2/frontend/src/layouts/AppShell.tsx
@@ -1,16 +1,16 @@
 import { useCallback, useState } from 'react'
 import { Outlet } from 'react-router-dom'
 import { Box, LinearProgress } from '@mui/material'
-import Navbar from '../../components/organisms/Navbar'
-import KpiRow from '../../components/molecules/KpiRow'
-import FilterBar from '../../components/molecules/FilterBar'
-import AssetTable from '../../components/organisms/AssetTable'
-import WorkspaceSplit from '../../components/organisms/WorkspaceSplit'
-import SettingsModal from '../../components/organisms/SettingsModal'
-import RebalanceDrawer from '../../components/organisms/RebalanceDrawer'
-import AssetProfileDrawer from '../../components/organisms/AssetProfileDrawer'
-import { PortfolioProvider, usePortfolioContext } from './PortfolioContext'
-import { useAppStore } from '../../store/useAppStore'
+import Navbar from '../components/organisms/Navbar'
+import KpiRow from '../components/molecules/KpiRow'
+import FilterBar from '../components/molecules/FilterBar'
+import AssetTable from '../components/organisms/AssetTable'
+import WorkspaceSplit from '../components/organisms/WorkspaceSplit'
+import SettingsModal from '../components/organisms/SettingsModal'
+import RebalanceDrawer from '../components/organisms/RebalanceDrawer'
+import AssetProfileDrawer from '../components/organisms/AssetProfileDrawer'
+import { PortfolioProvider, usePortfolioContext } from '../pages/portfolio/PortfolioContext'
+import { useAppStore } from '../store/useAppStore'
 
 function ShellInner({
   onSettingsOpen,
