+++ v2/frontend/src/pages/portfolio/tabs/AssetProfileTab.tsx
@@ -1,10 +1,7 @@
 import { Box, Typography, Stack } from '@mui/material'
 import { useAssetHistory } from '../../../hooks/useAssetHistory'
 import { useAppStore } from '../../../store/useAppStore'
-import AssetPriceChart from '../../../components/charts/AssetPriceChart'
-import AssetValueChart from '../../../components/charts/AssetValueChart'
-import AssetPnlChart from '../../../components/charts/AssetPnlChart'
-import AssetReturnChart from '../../../components/charts/AssetReturnChart'
+import { AssetPriceChart, AssetValueChart, AssetPnlChart, AssetReturnChart } from '../../../components/charts/asset'
 import Section from '../../../components/molecules/Section'
 import KpiGroup from '../../../components/molecules/KpiGroup'
 import { usePortfolioContext } from '../PortfolioContext'
