+++ v2/frontend/src/pages/portfolio/tabs/OpportunitiesTab.tsx
@@ -1,7 +1,5 @@
-import WinnersChart from '../../../components/charts/WinnersChart'
-import LosersChart from '../../../components/charts/LosersChart'
-import PositionPerformanceMap from '../../../components/charts/PositionPerformanceMap'
-import OpportunitiesChart from '../../../components/charts/OpportunitiesChart'
+import { WinnersChart, LosersChart, OpportunitiesChart } from '../../../components/charts/screener'
+import { PositionPerformanceMap } from '../../../components/charts/position'
 import { useAppStore } from '../../../store/useAppStore'
 import Section from '../../../components/molecules/Section'
 import { usePortfolioContext } from '../PortfolioContext'
